feat: Add xdg crate for config file discovery
The `config_path()` function has been updated to use `xdg` to find the
configuration file.

On non-Windows platforms: - It checks for the config file in
XDG_CONFIG_DIRS and $HOME/.config/jless/jless.toml. - If the config file
doesn't exist it falls back to $HOME/.jless.toml.

On Windows: - It uses dirs::config_dir() to find the config directory.

The default themes are defined.